Move decoder to separate file
